block: add uImage.FIT subimage block driver
Add a small block driver which exposes filesystem sub-images contained
in U-Boot uImage.FIT images as block devices.

The uImage.FIT image has to be stored directly on a block device or
partition, MTD device or partition, or UBI volume.

The driver is intended for systems using the U-Boot bootloader and
uses the root device hint left by the bootloader (or the user) in
the 'chosen' section of the device-tree.

Example:
/dts-v1/;
/ {
        chosen {
                rootdisk = <&mmc1 6>;
        };
};

Here partition 6 is used on a block device which is a descendant of the
device node referenced by &mmc1.

For MTD devices the mtdblock driver needs to be present. Use argument
-1 to indicate using the mtdblock device as lower device, e.g.:
    rootdisk = <&fit_part (-1)>;

For UBI volume 3 if the ubiblock driver needs to be present and the
ubiblock volume needs to be created via cmdline. Then it can be
referenced by the MTD device or partition hosting the UBI device and
the UBI volume ID as argument, e.g.:
    rootdisk = <&ubi_part 3>;

Signed-off-by: Daniel Golle <daniel@makrotopia.org>